Move oauth client code into client_oauth.go
